+++ doc/source/doxygen-docs/changelog.md
@@ -1,7 +1,9 @@
 \page changelog Change Log
 
 # Version 2.15.7: UNRELEASED
-(none yet)
+- Changes in libraries:
+  - \ref mrpt_io_grp
+    - Support for ZStd compression with mrpt::io::CCompressedInputStream and mrpt::io::CCompressedOutputStream
 
 # Version 2.15.6: Released Jan 29th, 2026
 - nanogui: Updated to new version (DockablePanel)